fhir_generator/requirements.txt
clickhouse_connect
fastapi==0.115.8
uvicorn==0.20.0
httpx==0.26.0
fhir-data-generator==1.8.0
circuit_breaker_fastapi
